test(TestUtil): extract support methods
